requirements.txt
joblib==1.2.0
mindspore==2.0.0
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.1
scikit_learn==1.3.0
tensorboardX==2.6
tensorboardX==2.6.2.2
torch==2.0.1+cu117
tqdm==4.65.0
